modified schaefer.py to just be for mintpy. mintpy results look good. in the future, can be integrated with my results?
