+++ src/py/methods/analyze_igram_simple.py
@@ -21,7 +21,7 @@ import sys
 import tqdm
 sys.path.append("..")
 
-from methods.schaefer import liu_deformation_from_alt, liu_alt_from_deformation, compute_bounding_box, plot_change
+from py.methods.run_mintpy_ts_analysis import liu_deformation_from_alt, liu_alt_from_deformation, compute_bounding_box, plot_change
 from utils import compute_stats, load_img, LatLon
 from data.consts import CALM_PROCESSSED_DATA_DIR, ISCE2_OUTPUTS_DIR
 from data.utils import get_date_for_alos
