+++ src/py/methods/resalt.py
@@ -6,9 +6,9 @@ from typing import List, Optional, Tuple
 import numpy as np
 import pandas as pd
 
-from methods.schaefer import get_norm_ddt
 from methods.soil_models import SoilMoistureModel
 from methods.simulate_sub_diff_solve import find_best_alt_diff
+from methods.run_analysis import get_norm_ddt
 
 
 class ReSALT_Type(enum.Enum):
